Sprint 13: JWT auth, routes, and Mongo setup complete

+++ controllers/clothingItems.js
@@ -1,25 +1,23 @@
-// controllers/clothingItems.js
 const mongoose = require('mongoose');
 const ClothingItem = require('../models/clothingItem');
-const { BAD_REQUEST, NOT_FOUND, INTERNAL_SERVER_ERROR } = require('../utils/errors');
+const { BAD_REQUEST, NOT_FOUND, FORBIDDEN, INTERNAL_SERVER_ERROR } = require('../utils/errors');
 
-// GET /items
 module.exports.getItems = async (req, res) => {
   try {
     const items = await ClothingItem.find({});
-    res.send(items);
+    return res.send(items);
   } catch (err) {
     console.error(err);
-    res.status(INTERNAL_SERVER_ERROR).send({ message: 'An error has occurred on the server.' });
+    return res
+      .status(INTERNAL_SERVER_ERROR)
+      .send({ message: 'An error has occurred on the server.' });
   }
 };
 
-// POST /items
 module.exports.createItem = async (req, res) => {
   try {
     const { name, weather, imageUrl } = req.body;
     const owner = req.user && req.user._id;
-
     const item = await ClothingItem.create({ name, weather, imageUrl, owner });
     return res.status(201).send(item);
   } catch (err) {
@@ -33,18 +31,22 @@ module.exports.createItem = async (req, res) => {
   }
 };
 
-// DELETE /items/:itemId
 module.exports.deleteItem = async (req, res) => {
   try {
     const { itemId } = req.params;
+    const userId = req.user && req.user._id;
 
     if (!mongoose.Types.ObjectId.isValid(itemId)) {
       return res.status(BAD_REQUEST).send({ message: 'Invalid item id' });
     }
 
     const item = await ClothingItem.findById(itemId).orFail();
-    await item.deleteOne();
 
+    if (item.owner.toString() !== userId) {
+      return res.status(FORBIDDEN).send({ message: 'Forbidden: not the owner' });
+    }
+
+    await item.deleteOne();
     return res.send({ message: 'Item deleted', item });
   } catch (err) {
     console.error(err);
@@ -60,7 +62,6 @@ module.exports.deleteItem = async (req, res) => {
   }
 };
 
-// PUT /items/:itemId/likes
 module.exports.likeItem = async (req, res) => {
   const { itemId } = req.params;
   const userId = req.user && req.user._id;
@@ -91,8 +92,7 @@ module.exports.likeItem = async (req, res) => {
   }
 };
 
-// DELETE /items/:itemId/likes
-module.exports.unlikeItem = async (req, res) => {
+module.exports.dislikeItem = async (req, res) => {
   const { itemId } = req.params;
   const userId = req.user && req.user._id;
